Template#apply() takes over View#applyTemplateToElement().
